requirements.txt
torch
torchvision
numpy
scikit-learn
timm
matplotlib
tqdm
jupyter
ipywidgets
flask
Pillow
